+++ src/settings.js
@@ -29,6 +29,15 @@ export function registerSettings() {
 		default: true,
 	})
 
+	game.settings.register(settingsKey, "showGMRulerToPlayers", {
+		name: "drag-ruler.settings.showGMRulerToPlayers.name",
+		hint: "drag-ruler.settings.showGMRulerToPlayers.hint",
+		scope: "world",
+		config: true,
+		type: Boolean,
+		default: true,
+	})
+
 	// This setting will be modified by the api if modules register to it
 	game.settings.register(settingsKey, "speedProvider", {
 		scope: "world",