+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## In development
 ### New features
 - If the [Terrain Ruler module](https://foundryvtt.com/packages/terrain-ruler/) is installed and activated, Drag Ruler will now take difficult terrain that was placed with the [TerrainLayer module](https://foundryvtt.com/packages/TerrainLayer/) into account.
+- The GM's Drag Ruler can now be hidden from non GM players via a setting.
 
 ## 1.3.4
 ### Module compatibility